pyproject.toml
[tool.poetry]
package-mode = false

[project]
name = "ai-stack"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name", email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0" 
dependencies = [
    "langchain (>=0.3.27,<0.4.0)",
    "langchain-ollama (>=0.3.6,<0.4.0)",
    "langgraph (>=0.6.5,<0.7.0)",
    "langchain-openai (>=0.3.31,<0.4.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "langgraph-checkpoint-sqlite (>=2.0.11,<3.0.0)",
    "langchain-community (>=0.3.27,<0.4.0)",
    "langchain-chroma (>=0.2.5,<0.3.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "pydantic (>=2.11.7,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
